refactor tests and add mock user data
