+++ packages/app/utils/mocks/Users.js
@@ -1,0 +1,51 @@
+const mongoose = require("mongoose");
+const bcrypt = require("bcrypt");
+const { UserType } = require("../../utils/constants");
+
+const mockUser = [
+  {
+    _id: new mongoose.Types.ObjectId(),
+    name: "John Doe",
+    email: "johndoe@example.com",
+    password: bcrypt.hashSync("password123", 10),
+    role: UserType.CUSTOMER,
+    is_verified: false,
+    createdAt: Date.now(),
+    updatedAt: Date.now(),
+  },
+  {
+    _id: new mongoose.Types.ObjectId(),
+    name: "John Doe",
+    email: "johndoe1@example.com",
+    password: bcrypt.hashSync("password123", 10),
+    role: UserType.CUSTOMER,
+    is_verified: true,
+    subscription_id: new mongoose.Types.ObjectId(),
+    createdAt: Date.now(),
+    updatedAt: Date.now(),
+  },
+  {
+    _id: new mongoose.Types.ObjectId(),
+    name: "John Doe",
+    email: "johndoe2@example.com",
+    password: bcrypt.hashSync("password123", 10),
+    role: UserType.ADMIN,
+    is_verified: true,
+    subscription_id: new mongoose.Types.ObjectId(),
+    createdAt: Date.now(),
+    updatedAt: Date.now(),
+  },
+  {
+    _id: new mongoose.Types.ObjectId(),
+    name: "John Doe",
+    email: "johndoe3@example.com",
+    password: bcrypt.hashSync("password123", 10),
+    role: UserType.OPERATOR,
+    is_verified: true,
+    subscription_id: new mongoose.Types.ObjectId(),
+    createdAt: Date.now(),
+    updatedAt: Date.now(),
+  },
+];
+
+module.exports = mockUser;
